remove amazon-vega-implementation
